Move BlockNumberAndHash to ckb_types
